// ec/src/models/api/msm/ect/partNumber/SearchPartNumberResponse$search.ts
export interface SearchPartNumberResponse$search extends MsmApiResponse {
	/**
	 * 総件数
	 * - 検索にヒットした総件数
	 * - example: 1022
	 * - field groups: @search
	 */
	totalCount: number;
	/**
	 * 型番確定フラグ
	 * - 型番が確定したかを表すフラグ
	 *   0:未確定
	 *   1:確定
	 * - example: 0
	 * - field groups: @default, @search, @detail
	 */
	completeFlag: Flag;
	/**
	 * ユニット品存在確定フラグ
	 * - ユニット品が存在し、一意に確定したかを表すフラグ
	 *   0:未確定
	 *   1:存在し、確定(検索結果の型番すべてが同一のユニット品)
	 * - example: 0
	 * - field groups: @default, @search, @detail
	 */
	unitProductFlag: Flag;
	/**
	 * スペック項目リスト
	 * - 型番リスト、指定中の仕様・寸法情報、類似品検索で使用するスペック項目のリスト
	 * - field groups: @search, @detail
	 */
	specList?: Spec[];
	/**
	 * 規制項目リスト
	 * - 貿易コンプラから連携された商品の規制情報
	 * - field groups: @search, @detail
	 * - NOTE: シリーズ内で使用されている項目のみ返却
	 */
	regulationList?: Regulation[];
	/**
	 * 型番リスト
	 * - field groups: @default, @search, @detail
	 */
	partNumberList: PartNumber[];
	/**
	 * CAD IDリスト
	 * - 絞りこまれた型番に紐づくCAD情報のIDのリスト。
	 *   CADデータ取得時、プレビュー時に指定する。
	 * - example: [10000000449,10000000459]
	 * - field groups: @search
	 */
	cadIdList?: string[];
	/**
	 * 単純品スペック項目コードリスト
	 * - 単純品テンプレート時に使用するスペック項目を指定するためのリスト。
	 *   見出しは「型番絞り込み候補情報.仕様・寸法リスト」の該当するコードのスペック名およびスペック値(ブルダウン内)を使用し、各型番のスペック値は「型番情報.型番リスト.スペック値リスト」の該当するコードの値を使用する
	 * - example: [C001,C003,D001,C004]
	 * - field groups: @search, @detail
	 */
	simpleProductSpecCodeList?: string[];
	/**
	 * 進捗ガイド数
	 * - 矢羽根の母数
	 * - example: 10
	 * - field groups: @search
	 */
	maxGuideCount: number;
	/**
	 * 進捗状況数
	 * - 絞込み状況を表す矢羽根数
	 * - example: 5
	 * - field groups: @search
	 */
	guideCount: number;
	/**
	 * ユニットコードリスト
	 * - コンベヤの部材品を特定するためのユニットコード
	 * - example: 018256445568
	 * - field groups: @default, @search, @detail
	 */
	unitCodeList?: string[];
	/**
	 * 在庫品情報
	 * - 在庫品情報
	 * - field groups: @search
	 * - NOTE: 別紙３参照
	 */
	stockItem: StockItem;
	/**
	 * 仕様・寸法リスト
	 * - 選択項目の一覧を表示
	 * - field groups: @search
	 */
	partNumberSpecList: PartNumberSpec[];
	/**
	 * グループ項目リスト
	 * - 並列で表示するスペック項目のリスト
	 * - field groups: @search
	 */
	groupSpecList: GroupSpec[];
	/**
	 * 追加工注意喚起文
	 * - 追加工領域の上部に表示される注意喚起文
	 * - field groups: @search
	 */
	alterationNoticeText?: string;
	/**
	 * 追加工スペックリスト
	 * - 追加工が存在する場合に一覧を表示
	 * - field groups: @search
	 */
	alterationSpecList: AlterationSpec[];
	/**
	 * CADタイプリスト
	 * - CADの選択状態
	 * - example: 1,2
	 * - field groups: @search
	 */
	cadTypeList: CadType[];
	/**
	 * 出荷日数リスト
	 * - 出荷日の選択状態
	 * - field groups: @search
	 */
	daysToShipList: DaysToShip[];
	/**
	 * ドリルダウンスペックリスト
	 * - ドリルダウンで表示するスペックのリスト
	 * - field groups: @search
	 */
	drillDownSpecList?: DrillDownSpec[];
	/**
	 * スペックグループリスト
	 * - スペックグループのリスト
	 * - field groups: @search
	 */
	partNumberSpecGroupList?: PartNumberSpecGroup[];
	/**
	 * 追加工スペックグループリスト
	 * - 追加工スペックグループのリスト
	 * - field groups: @search
	 */
	alterationSpecGroupList?: AlterationSpecGroup[];
	/**
	 * 通貨コード
	 * - 通貨コード
	 *   JPY: 日本円
	 *   RMB: 人民元
	 * - example: JPY
	 * - field groups: @search, @detail
	 */
	currencyCode?: string;
}
// ...
export interface Regulation {
	/**
	 * 規制コード
	 * - 規制情報のコード値
	 * - field groups: @search, @detail
	 */
	regulationCode: string;
	/**
	 * 規制名
	 * - 規制情報の項目名
	 * - field groups: @search, @detail
	 */
	regulationName: string;
}
// ...
export interface StockItem {
	/**
	 * 在庫品フラグ
	 * - 在庫品対象の指定
	 *   0: 指定しない
	 *   1: 指定する
	 * - field groups: @search
	 * - NOTE: 現在の条件に該当する在庫品が１件でもあれば"1"、無ければ"0"
	 */
	stockItemFlag: Flag;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 * - NOTE: 現在の条件に該当する在庫品が１件でもあれば"0"、無ければ"1"
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みの選択肢の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 * - NOTE: リクエストの内容に応じて返却
	 */
	selectedFlag: Flag;
}

/** 仕様・寸法 */
export interface PartNumberSpec {
	/**
	 * スペック項目コード
	 * - スペック項目のコード
	 * - example: E01
	 * - field groups: @search
	 */
	specCode: string;
	/**
	 * スペック項目名
	 * - スペック項目名
	 * - example: 長さ
	 * - field groups: @search
	 */
	specName: string;
	/**
	 * スペック項目単位
	 * - スペック項目の単位
	 * - example: mm
	 * - field groups: @search
	 */
	specUnit?: string;
	/**
	 * スペック項目表示タイプ
	 * - スペックの表示種別
	 *   1: テキストボタン形式
	 *   2: テキスト選択形式（1列）
	 *   3: テキスト選択形式（2列）
	 *   4: テキスト選択形式（3列）
	 *   5: 画像ボタン形式（1列）
	 *   6: 画像ボタン形式（2列）
	 *   7: 画像ボタン形式（3列）
	 *   8: 数値入力形式
	 *   9: リスト選択形式
	 *   10: ツリー(入れ子)選択形式
	 * - example: 1
	 * - field groups: @search
	 * - NOTE: 型番検索APIの仕様・寸法リストでは
	 *          11: 数値集約形式
	 *          12: 中央表示
	 *         は返却しない
	 *         型番検索APIの仕様・寸法リストでは
	 *         13～21の表示タイプは返却しない、該当するタイプは通常テンプレート向けに読み替える。（詳しい内容は別紙４参照）
	 */
	specViewType: SpecViewType;
	/**
	 * 開閉タイプ
	 * - スペック表示の開閉状態
	 *   1: 開
	 *   2: 閉
	 *   3: 開閉ボタン非表示
	 * - example: 1
	 * - field groups: @search
	 */
	openCloseType?: OpenCloseType;
	/**
	 * 補足表示タイプ
	 * - スペックの補足表示種別
	 *   1: 通常
	 *   2: 外形図
	 *   3: 詳細
	 *   4: 拡大
	 *   5: イラスト
	 * - example: 2
	 * - field groups: @search
	 * - NOTE: 4,5の場合はスペック値画像URLの画像を用いる
	 */
	supplementType?: SupplementType;
	/**
	 * 外形図画像URL
	 * - 外形図ダイアログで表示する画像のURL
	 * - example: //xxx/illustration/press/P010105_C3_F.png
	 * - field groups: @search
	 */
	specImageUrl?: string;
	/**
	 * 詳細HTML
	 * - 詳細ダイアログで表示するHTML
	 * - field groups: @search
	 */
	detailHtml?: string;
	/**
	 * スペック説明画像URL
	 * - スペックの説明画像のURL
	 * - field groups: @search
	 */
	specDescriptionImageUrl?: string;
	/**
	 * スペック注意喚起文
	 * - 注意喚起文の本文
	 * - field groups: @search
	 */
	specNoticeText?: string;
	/**
	 * 数値入力項目
	 * - 数値入力項目
	 * - field groups: @search
	 */
	numericSpec?: NumericSpec;
	/**
	 * スペック値リスト
	 * - field groups: @search
	 * - NOTE: リスト内のすべてのスペック値が非表示の場合は、リスト(ボタン)を非表示
	 */
	specValueList: SpecValue[];
	/**
	 * 表示順
	 * - スペック項目の表示順
	 * - example: 1
	 * - field groups: @search
	 */
	displayOrder?: number;
}

/** 数値入力項目 */
export interface NumericSpec {
	/**
	 * 非表示フラグ
	 * - 数値入力項目を表示するかどうか
	 *   0: 表示する
	 *   1: 表示しない
	 * - field groups: @search
	 * - NOTE: そもそも数値入力欄が存在しない場合は、「数値入力項目」自体を返却しない
	 *         (空の値"{}"を返却)
	 */
	hiddenFlag: Flag;
	/**
	 * スペック値
	 * - 数値入力形式のテキストフィールドに表示する値(リクエスト値)
	 * - example: 100
	 * - field groups: @search
	 */
	specValue?: string;
	/**
	 * 数値範囲リスト
	 * - 数値範囲の選択範囲
	 * - field groups: @search
	 */
	specValueRangeList: SpecValueRange[];
}

/** 数値範囲 */
export interface SpecValueRange {
	/**
	 * 最小値
	 * - 指定可能な最小値
	 * - example: 1.0
	 * - field groups: @search
	 */
	minValue?: number;
	/**
	 * 最大値
	 * - 指定可能な最大値
	 * - example: 9.5
	 * - field groups: @search
	 */
	maxValue?: number;
	/**
	 * 刻み値
	 * - 値の刻み幅
	 * - example: 0.5
	 * - field groups: @search
	 */
	stepValue?: number;
}

/** スペック値 */
export interface SpecValue {
	/**
	 * スペック値
	 * - スペック値(コードもしくは数値)
	 * - example: 1
	 * - field groups: @search
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 鉄
	 * - field groups: @search
	 */
	specValueDisp: string;
	/**
	 * スペック値画像URL
	 * - スペック値の画像URL
	 * - field groups: @search
	 * - NOTE: 設計書では必須
	 */
	specValueImageUrl?: string;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みのスペック値の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
	/**
	 * 子スペック値リスト
	 * - 2段表示のスペック値リスト
	 * - field groups: @search
	 */
	childSpecValueList?: ChildSpecValue[];
	/**
	 * スペック値属性リスト
	 * - スペック値の属性情報
	 *   PU向けのスペック表示(表形式)に対応している場合に設定
	 * - field groups: @search
	 */
	specValueAttributeList?: SpecValueAttribute[];
}

/** 子スペック値 */
export interface ChildSpecValue {
	/**
	 * スペック値
	 * - スペック値(コードもしくは数値)
	 * - example: 1
	 * - field groups: @search
	 */
	specValue: string;
	/**
	 * スペック値表示文言
	 * - スペック値の表示文言
	 * - example: 鉄
	 * - field groups: @search
	 */
	specValueDisp: string;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みのスペック値の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
}

/** スペック値属性 */
export interface SpecValueAttribute {
	/**
	 * 属性リスト
	 * - 属性のリスト
	 * - field groups: @search
	 */
	attributeList?: Attribute[];
}
// ...
export interface AlterationSpec {
	/**
	 * スペック項目コード
	 * - スペック項目コード
	 * - example: E01
	 * - field groups: @search
	 */
	specCode: string;
	/**
	 * CAD用スペック項目コード
	 * - CAD用のスペック項目コード
	 * - example: LKC
	 * - field groups: @search
	 * - NOTE: SINUSでのCADデータ生成時に使用
	 */
	cadSpecCode?: string;
	/**
	 * スペック項目名
	 * - スペック項目名
	 * - example: 材質
	 * - field groups: @search
	 */
	specName: string;
	/**
	 * スペック項目単位
	 * - スペック項目の単位
	 * - example: mm
	 * - field groups: @search
	 */
	specUnit?: string;
	/**
	 * スペック項目表示タイプ
	 * - スペックの表示種別
	 *   2: テキスト選択形式（1列）
	 *   3: テキスト選択形式（2列）
	 *   4: テキスト選択形式（3列）
	 *   8: 数値入力形式
	 *   9: リスト選択形式
	 * - example: 8
	 * - field groups: @search
	 * - NOTE: 型番検索APIの追加工スペックリストでは
	 *           1: テキストボタン形式
	 *           5: 画像ボタン形式（1列）
	 *           6: 画像ボタン形式（2列）
	 *           7: 画像ボタン形式（3列）
	 *          10: ツリー(入れ子)選択形式
	 *          11: 数値集約形式
	 *          12: 中央表示
	 *         は返却しない
	 */
	specViewType: AlterationSpecViewType;
	/**
	 * 補足表示タイプ
	 * - スペックの補足表示種別
	 *   1: 通常
	 *   2: 外形図
	 *   3: 詳細
	 *   4: 拡大
	 *   5: イラスト
	 * - example: 2
	 * - field groups: @search
	 * - NOTE: 4,5の場合はスペック値画像URLの画像を用いる
	 */
	supplementType: SupplementType;
	/**
	 * 外形図画像URL
	 * - 外形図ダイアログで表示する画像のURL
	 * - example: //xxx/illustration/press/P010105_C3_F.png
	 * - field groups: @search
	 */
	specImageUrl?: string;
	/**
	 * 詳細HTML
	 * - 詳細ダイアログで表示するHTML
	 * - field groups: @search
	 */
	detailHtml?: string;
	/**
	 * スペック説明画像URL
	 * - スペックの説明画像のURL
	 * - field groups: @search
	 */
	specDescriptionImageUrl?: string;
	/**
	 * スペック注意喚起文
	 * - 注意喚起文の本文
	 * - field groups: @search
	 */
	specNoticeText?: string;
	/**
	 * 非表示フラグ
	 * - 規格式結果で隠すべきときで、ユーザに選択されていなければ隠す
	 *   0: 表示
	 *   1: 非表示
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 範囲値が埋まっていたり選択値がひとつでも選択されていた場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - field groups: @search
	 */
	selectedFlag: Flag;
	/**
	 * 数値入力項目
	 * - 数値入力項目
	 * - field groups: @search
	 */
	numericSpec: AlterationNumericSpec;
	/**
	 * スペック値リスト
	 * - field groups: @search
	 */
	specValueList: AlterationSpecValue[];
}

/** 数値入力項目 */
export interface AlterationNumericSpec {
	/**
	 * 非表示フラグ
	 * - 数値入力項目を表示するかどうか
	 *   0: 表示する
	 *   1: 表示しない
	 * - field groups: @search
	 * - NOTE: そもそも数値入力欄が存在しない場合は、「数値入力項目」自体を返却しない
	 *         (空の値"{}"を返却)
	 */
	hiddenFlag?: Flag;
	/**
	 * スペック値
	 * - 数値入力形式のテキストフィールドに表示する値(リクエスト値)
	 * - example: 100
	 * - field groups: @search
	 */
	specValue?: string;
	/**
	 * 数値範囲リスト
	 * - 数値範囲の選択範囲
	 * - field groups: @search
	 */
	specValueRangeList: SpecValueRange[];
}
// ...
export interface CadType {
	/**
	 * CADタイプ
	 * - CADが存在する場合に表示
	 *   1: 2D
	 *   2: 3D
	 * - field groups: @search
	 */
	cadType: string;
	/**
	 * CADタイプ表示文言
	 * - CADタイプの表示文言
	 * - field groups: @search
	 */
	cadTypeDisp: string;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みの選択肢の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
}

/** 出荷日数 */
export interface DaysToShip {
	/**
	 * 出荷日数
	 * - 出荷までの実働日数
	 *   -1: 即納
	 *   0: 当日出荷
	 *   1～98: 出荷日数
	 * - maxLength: 3
	 * - example: 1
	 * - field groups: @search
	 */
	daysToShip: number;
	/**
	 * 非表示フラグ
	 * - field groups: @search
	 */
	hiddenFlag?: Flag;
	/**
	 * 選択済みフラグ
	 * - 選択済みの選択肢の場合に設定
	 *   0: 未選択
	 *   1: 選択済
	 * - example: 0
	 * - field groups: @search
	 */
	selectedFlag: Flag;
}
// ...
export const SpecViewType = {
	TEXT_BUTTON: '1',
	TEXT_SINGLE_LINE: '2',
	TEXT_DOUBLE_LINE: '3',
	TEXT_TRIPLE_LINE: '4',
	IMAGE_SINGLE_LINE: '5',
	IMAGE_DOUBLE_LINE: '6',
	IMAGE_TRIPLE_LINE: '7',
	NUMERIC: '8',
	LIST: '9',
	TREE: '10',
} as const;
export type SpecViewType = typeof SpecViewType[keyof typeof SpecViewType];

/**
 * スペック項目表示タイプ
 * - スペックの表示種別
 *   2: テキスト選択形式（1列）
 *   3: テキスト選択形式（2列）
 *   4: テキスト選択形式（3列）
 *   8: 数値入力形式
 *   9: リスト選択形式
 */
export type AlterationSpecViewType = Extract<
	SpecViewType,
	'2' | '3' | '4' | '8' | '9'
>;

/**
 * 関連情報
 *   1: SDS(MSDS)
 *   2: データシート
 *   3: ミスミ定期便申し込み
 *   4: サンプル品提供申し込み
 */
export const RelatedLinkType = {
	MSDS: '1',
	DATA_SHEET: '2',
	MISUMI_REGULAR_SERVICE: '3',
	REQUEST_SAMPLE: '4',
} as const;
export type RelatedLinkType =
	typeof RelatedLinkType[keyof typeof RelatedLinkType];

/**
 * 補足表示タイプ
 * - スペックの補足表示種別
 *   1: 通常
 *   2: 外形図
 *   3: 詳細
 *   4: 拡大
 *   5: イラスト
 */
export const SupplementType = {
	NORMAL: '1',
	DRAWING: '2',
	DETAIL: '3',
	ZOOM_IMAGE: '4',
	ILLUSTRATION: '5',
} as const;
export type SupplementType = typeof SupplementType[keyof typeof SupplementType];